refactor: move InterruptibleParserATNSimulator into pmd-kotlin as package-private
Use Thread.currentThread().isInterrupted() to avoid clearing the interrupted flag.
General pmd-core solution tracked in pmd/pmd#6684.
